refactor test classes making http requests; add one more api test
